Added several elements.
Added an editable Table, with full editing and add & remove rows & columns functionality, with header and footer support as well. Credit goes to https://github.com/nlulic/slate-table
Also added a Table of Contents & Page Break component, both of which have no function as of yet. Said functionality will be added when the document is to be converted to a PDF file.